UPSTREAM: arm64: ptrace: fix compat registes get/set to be endian clean
(cherry picked from commit 6a2e5e521c333a0b56cb60dc5587e3f90859c5e7)

On a BE system the wrong half of the X registers is retrieved/written
when attempting to get/set the value of aarch32 registers through
ptrace.

Ensure that types are the correct width so that the relevant
casting occurs.

Change-Id: I83597e0ce55033b2c485de68ed2ba8af625d972f
Signed-off-by: Matthew Leach <matthew.leach@arm.com>
Reviewed-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>